+++ src/translations/en-US/calendar.php
@@ -130,6 +130,19 @@ $events = [
     'Th' => 'Th',
     'Fr' => 'Fr',
     'Sa' => 'Sa',
+
+    'January'   => 'January',
+    'February'  => 'February',
+    'March'     => 'March',
+    'April'     => 'April',
+    'May'       => 'May',
+    'June'      => 'June',
+    'July'      => 'July',
+    'August'    => 'August',
+    'September' => 'September',
+    'October'   => 'October',
+    'November'  => 'November',
+    'December'  => 'December',
 ];
 
 $calendars = [